# api/openapi/identity-service.v1.yaml
openapi: 3.0.3
info:
  title: identity-service API
  version: 1.0.0
  description: |
    Contract for the go-ory-auth-example identity-service (ADR-0010).
    Human-readable guide: docs/api/identity-service-api.md.

    Credential is bound to plane:
    - `/v1/*` accepts only `Authorization: Bearer <kratos session token>` (mobile).
    - `/admin/v1/*` accepts only the Kratos session cookie (admin web).

    Errors use RFC 9457 Problem Details (`application/problem+json`) with a stable `code`.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
  /v1/me:
    get:
      tags: [me]
      operationId: getMe
      summary: Current customer profile
      security: [{ bearerSession: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [me]
      operationId: updateMe
      summary: Update the current customer's profile (requires verified email)
      security: [{ bearerSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /admin/v1/me:
    get:
      tags: [admin]
      operationId: getAdminMe
      summary: Current admin session; the only admin endpoint reachable at AAL1
      description: |
        403 codes drive the admin web routing: `not_admin`, `aal2_required`,
        `mfa_enrollment_required`.
      security: [{ cookieSession: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminMe" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /admin/v1/customers:
    get:
      tags: [customers]
      operationId: listCustomers
      summary: List customers (permission view_customers)
      security: [{ cookieSession: [] }]
      parameters:
        - { name: email, in: query, schema: { type: string, format: email } }
        - { name: state, in: query, schema: { $ref: "#/components/schemas/IdentityState" } }
        - { $ref: "#/components/parameters/PageSize" }
        - { $ref: "#/components/parameters/PageToken" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerPage" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
...
  /admin/v1/customers/{id}:
    parameters: [{ $ref: "#/components/parameters/IdentityId" }]
    get:
      tags: [customers]
      operationId: getCustomer
      summary: Get one customer (permission view_customers)
      security: [{ cookieSession: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/v1/customers/{id}/disable:
    parameters: [{ $ref: "#/components/parameters/IdentityId" }]
    post:
      tags: [customers]
      operationId: disableCustomer
      summary: Disable a customer and revoke all sessions (permission manage_customers; idempotent)
      security: [{ cookieSession: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerState" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/v1/customers/{id}/enable:
    parameters: [{ $ref: "#/components/parameters/IdentityId" }]
    post:
      tags: [customers]
      operationId: enableCustomer
      summary: Re-enable a customer (permission manage_customers; idempotent)
      security: [{ cookieSession: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerState" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/v1/customers/{id}/sessions:
    parameters: [{ $ref: "#/components/parameters/IdentityId" }]
    delete:
      tags: [customers]
      operationId: revokeCustomerSessions
      summary: Revoke all sessions of a customer (permission manage_customers)
      security: [{ cookieSession: [] }]
      responses:
        "204": { description: Revoked }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/v1/admins:
    get:
      tags: [admins]
      operationId: listAdmins
      summary: List admins (permission manage_admins)
      security: [{ cookieSession: [] }]
      parameters:
        - { $ref: "#/components/parameters/PageSize" }
        - { $ref: "#/components/parameters/PageToken" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPage" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [admins]
      operationId: inviteAdmin
      summary: Invite an admin (permission manage_admins). The invitation link is emailed, never returned.
      security: [{ cookieSession: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteAdminRequest" }
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitedAdmin" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /admin/v1/admins/{id}/role:
    parameters: [{ $ref: "#/components/parameters/IdentityId" }]
    put:
      tags: [admins]
      operationId: changeAdminRole
      summary: Change an admin's role (permission manage_admins; not self; keeps >= 1 super_admin)
      security: [{ cookieSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeRoleRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Admin" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /admin/v1/audit-events:
    get:
      tags: [audit]
      operationId: listAuditEvents
      summary: List audit events, newest first (permission view_audit)
      security: [{ cookieSession: [] }]
      parameters:
        - { name: target_type, in: query, schema: { type: string } }
        - { name: target_id, in: query, schema: { type: string } }
        - { name: actor_id, in: query, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/PageSize" }
        - { $ref: "#/components/parameters/PageToken" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditEventPage" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }

  # Served on the webhook port (:8081) only; listed for completeness.
...
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      description: Kratos session token from a native (API) flow.
    cookieSession:
      type: apiKey
      in: cookie
      name: ory_kratos_session
    webhookKey:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    IdentityId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    PageToken:
      name: page_token
      in: query
      schema: { type: string }
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, example: "https://docs.go-ory-auth-example.local/problems/aal2-required" }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: |
            Stable machine code. Known values: unauthenticated, forbidden, not_admin,
            aal2_required, mfa_enrollment_required, email_not_verified, not_found,
            conflict, validation_failed, rate_limited, dependency_unavailable, internal.
            Clients must tolerate unknown values.
        detail: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, example: too_long }
    PersonName:
      type: object
      properties:
        first: { type: string, maxLength: 100 }
        last: { type: string, maxLength: 100 }
    IdentityState:
      type: string
      enum: [active, inactive]
    Role:
      type: string
      enum: [support, admin, super_admin]
    Permission:
      type: string
      enum: [view_customers, manage_customers, manage_admins, view_audit, reveal_customer_pii]
    Me:
      type: object
      required: [id, email, email_verified, locale, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        email_verified: { type: boolean }
        name: { $ref: "#/components/schemas/PersonName" }
        display_name: { type: string, nullable: true, maxLength: 100 }
        avatar_url: { type: string, nullable: true, maxLength: 2048 }
        locale: { type: string, example: vi-VN }
        created_at: { type: string, format: date-time }
    UpdateMeRequest:
      type: object
      additionalProperties: false
      properties:
        display_name: { type: string, nullable: true, maxLength: 100 }
        avatar_url: { type: string, nullable: true, maxLength: 2048 }
        locale: { type: string, maxLength: 35 }
...
    AdminMe:
      type: object
      required: [id, email, aal, roles, permissions, session_expires_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { $ref: "#/components/schemas/PersonName" }
        aal: { type: string, enum: [aal1, aal2] }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
        permissions:
          type: array
          items: { $ref: "#/components/schemas/Permission" }
        session_expires_at: { type: string, format: date-time }
    Customer:
      type: object
      required: [id, email, email_verified, state, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        email_verified: { type: boolean }
        name: { $ref: "#/components/schemas/PersonName" }
        state: { $ref: "#/components/schemas/IdentityState" }
        display_name: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    CustomerPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Customer" }
        next_page_token: { type: string }
    CustomerState:
      type: object
      required: [id, state]
      properties:
        id: { type: string, format: uuid }
        state: { $ref: "#/components/schemas/IdentityState" }
    ReasonRequest:
      type: object
      properties:
        reason: { type: string, maxLength: 500 }
    Admin:
      type: object
      required: [id, email, role, state, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { $ref: "#/components/schemas/PersonName" }
        role: { $ref: "#/components/schemas/Role" }
        state: { $ref: "#/components/schemas/IdentityState" }
        mfa_enrolled: { type: boolean }
        created_at: { type: string, format: date-time }
    AdminPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Admin" }
        next_page_token: { type: string }
    InviteAdminRequest:
      type: object
      required: [email, role]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        name: { $ref: "#/components/schemas/PersonName" }
        role: { $ref: "#/components/schemas/Role" }
    InvitedAdmin:
      type: object
      required: [id, email, role, invitation_expires_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        invitation_expires_at: { type: string, format: date-time }
    ChangeRoleRequest:
      type: object
      required: [role]
      additionalProperties: false
      properties:
        role: { $ref: "#/components/schemas/Role" }
    AuditEvent:
      type: object
      required: [id, occurred_at, actor_id, action, target_type, target_id, request_id, details]
      properties:
        id: { type: integer, format: int64 }
        occurred_at: { type: string, format: date-time }
        actor_id: { type: string, format: uuid }
        action: { type: string, example: customer.disabled }
        target_type: { type: string, example: customer }
        target_id: { type: string }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
    AuditEventPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AuditEvent" }
        next_page_token: { type: string }
...